user activity now gets persisted periodically. Improved caching.  And other minor changes

+++ src/main/java/nl/nielsvanbruggen/videostreamingplatform/watched/WatchedDTO.java
@@ -7,7 +7,7 @@ public record WatchedDTO(
         String name,
         int index,
         int season,
-        float timestamp,
+        double timestamp,
         long mediaId,
         Instant updatedAt,
         double duration
